requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# AI and MCP dependencies
openai==1.3.0
anthropic==0.7.0
mcp==0.4.0
